lambda/requirements.txt
# boto3==1.16.37
# aws-lambda-powertools==1.10.2
mysql-connector-python==8.0.27
# PyMySQL==1.0.2